+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * Export sorted by DisplayName for better comparison
   * Enable Filter property to be used on export
     FIXES [#3919](https://github.com/microsoft/Microsoft365DSC/issues/3919)
+
 # 1.23.1115.1
 
 * AADApplication