+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # UNRELEASED
 
 * DEPENDENCIES
+  * Updated Microsoft.Graph to version 2.10.0.
   * Updated MSCloudLoginAssistant to version 1.1.0.
 
 # 1.23.1122.1