+++ CHANGELOG.md
@@ -11,6 +11,9 @@
   IntuneWindowsAutopilotDeploymentProfileAzureADJoined
   * Removed Id and all authentication parameters from PSBoundParameters in Test-TargetResource
   FIXES [#3888](https://github.com/microsoft/Microsoft365DSC/issues/3888)
+* DEPENDENCIES
+  * Updated Microsoft.Graph to version 2.10.0.
+  * Updated MSCloudLoginAssistant to version 1.1.0.
 
 # 1.23.1122.1